Organize test files and notes into proper directories (Issue #7)
Major changes:
- Created /tests/ directory and moved all test files with clear naming (test-demo##-*.mjs)
- Consolidated notes and documentation into /notes/ directory
- Updated package.json with new test file paths
- Fixed all import statements in test files to reference correct demo directories
- Removed temporary debugging scripts (test-parry-consistency.sh, find-failing-test.sh, find-all-failures.sh)

Test files reorganized:
- 17 demo test files moved from demos/* to tests/test-demo##-*.mjs
- 3 demo15 chatbot tests organized (ELIZA, PARRY, ALICE)
- Verification scripts moved to tests/

Notes/documentation reorganized:
- 7 documentation files moved to notes/ with demo15- prefix
- Integration reports, summaries, and test results now in notes/

All tests verified working from new locations:
- Demo 01: 144/144 tests passing
- Demo 02: 141/141 tests passing
- Updated relative paths in test files to point to ../demos/##-*/

This completes the repository reorganization requested in Issue #7.
