+++ notes/demo15-eliza-fixes-integration.md
