+++ package.json
@@ -6,24 +6,24 @@
   "scripts": {
     "test": "npm run test:all",
     "test:all": "npm run test:demo01 && npm run test:demo02 && npm run test:demo03 && npm run test:demo04 && npm run test:demo05 && npm run test:demo06 && (npm run test:demo07 || echo 'Demo 07 failed (known memory issue)') && npm run test:demo08 && npm run test:demo09 && npm run test:demo10 && npm run test:demo11 && npm run test:demo12 && npm run test:demo13 && npm run test:demo14 && npm run test:demo15",
-    "test:demo01": "node demos/01-eliza/test-eliza.mjs",
-    "test:demo02": "node demos/02-tokenization/test-tokenization.mjs",
-    "test:demo03": "node demos/03-embeddings/test-clustering.mjs",
-    "test:demo04": "node demos/04-attention/test-attention.mjs",
-    "test:demo05": "node demos/05-transformer/test-transformer.mjs",
-    "test:demo06": "node demos/06-gpt-playground/test-sampling.mjs",
-    "test:demo07": "node demos/07-rag/test-rag.mjs",
-    "test:demo08": "node demos/08-topic-modeling/test-lda.mjs",
-    "test:demo09": "node demos/09-sentiment/test-sentiment.mjs",
-    "test:demo10": "node demos/10-pos-tagging/test-pos.mjs",
-    "test:demo11": "node demos/11-analogies/test-analogies.mjs",
-    "test:demo12": "node demos/12-semantic-search/test-bm25.mjs",
-    "test:demo13": "node demos/13-bert-mlm/test-bert.mjs",
-    "test:demo14": "node demos/14-embeddings-comparison/test-comparison.mjs",
+    "test:demo01": "node tests/test-demo01-eliza.mjs",
+    "test:demo02": "node tests/test-demo02-tokenization.mjs",
+    "test:demo03": "node tests/test-demo03-clustering.mjs",
+    "test:demo04": "node tests/test-demo04-attention.mjs",
+    "test:demo05": "node tests/test-demo05-transformer.mjs",
+    "test:demo06": "node tests/test-demo06-sampling.mjs",
+    "test:demo07": "node tests/test-demo07-rag.mjs",
+    "test:demo08": "node tests/test-demo08-lda.mjs",
+    "test:demo09": "node tests/test-demo09-sentiment.mjs",
+    "test:demo10": "node tests/test-demo10-pos.mjs",
+    "test:demo11": "node tests/test-demo11-analogies.mjs",
+    "test:demo12": "node tests/test-demo12-bm25.mjs",
+    "test:demo13": "node tests/test-demo13-bert.mjs",
+    "test:demo14": "node tests/test-demo14-comparison.mjs",
     "test:demo15": "npm run test:demo15:eliza && npm run test:demo15:parry && npm run test:demo15:alice",
-    "test:demo15:eliza": "node demos/15-chatbot-evolution/test-eliza-fixes.mjs",
-    "test:demo15:parry": "node demos/15-chatbot-evolution/test-parry-final.mjs",
-    "test:demo15:alice": "node demos/15-chatbot-evolution/test-alice-full-node.mjs",
+    "test:demo15:eliza": "node tests/test-demo15-eliza-fixes.mjs",
+    "test:demo15:parry": "node tests/test-demo15-parry-final.mjs",
+    "test:demo15:alice": "node tests/test-demo15-alice-full.mjs",
     "test:chatbots": "npm run test:demo15",
     "test:eliza": "npm run test:demo01 && npm run test:demo15:eliza",
     "test:tokenization": "npm run test:demo02",
