+++ tests/test-demo04-attention.mjs
