+++ tests/test-demo09-sentiment.mjs
