+++ tests/test-demo15-parry-final.mjs
@@ -1,4 +1,4 @@
-import { Parry } from './js/parry.js';
+import { Parry } from '../demos/15-chatbot-evolution/js/parry.js';
 
 console.log('=== COMPREHENSIVE TESTING AFTER FIXES ===\n');
